+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Buffer IMAP client writes #3888
 - move `DC_CHAT_ID_ARCHIVED_LINK` to the top of chat lists
   and make `dc_get_fresh_msg_cnt()` work for `DC_CHAT_ID_ARCHIVED_LINK` #3918
+- Update provider database
 
 ### API-Changes
 - jsonrpc: add python API for webxdc updates #3872